[0.0.4] Use a rollover offset category instead of Inflows
This commit uses a completely separate category for the positive, balance offset transaction. This also applies a negative budgeted amount to that category to correct the to be budgeted (inflows) balance. This should be simpler than the previous flow. This also adds prettier to make our code prettier and makes sure it is ignoring based on the `.gitignore` using the ignore path param.