[0.1.0] Rewrite the rollover logic, use server knowledge
This commit completely rewrites the rollover logic to use the latest server knowledge, as well as the get all months, to know when something has been updated and reduce the number of calls made per run. This also uses the rollover offset budget as well as `GROUPS_TO_OFFSET` to offset spending in a non-budgeted category (categories that are paid for by external sources, like a business category or something).