[1.0.0] Add support for future budgeting
This commit adds support for budgeting in the next month in the future. It rolls over the categories in the groups to offset by assigning in the future month, then it offsets that amount in a new budget category called "Future budgeted". It will reset the budgeted offset when the new month starts (and use a transaction like normal).