[1.2.1] Stop counting positive in unbudgeted as available
Prior to this, a positive balance in the unbudgeted categories was counted towards the available to budget amounts, so money that was owed to someone could potentially be budgeted away and no longer present. This commit fixes that—a negative balance takes away from available, but a positive balance does not add to it.